backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
ecdsa==0.19.0
Faker==37.0.0
fastapi==0.110.0
Flask==3.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.1
MarkupSafe==3.0.2
numpy==1.26.4
ollama==0.4.7
openai==1.65.2
packaging==24.2
pandas==2.2.1
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
PyMySQL==1.1.1
pypdf==5.3.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.1
requests==2.32.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.27.1
Werkzeug==3.1.3
wheel==0.45.1
Whoosh==2.7.4
